feat: Add cart animation on 'Add to Cart' button click
This commit introduces a new function, animateCart(), to handle the animation of the cart when the user clicks the 'Add to Cart' button. The animation class 'animate-cart' is applied to the cart element, and it is removed after a millisecond delay using setTimeout.

Additionally, the "addToCart" click event listener now includes animateCart() added onto addToCartHandler()

Signed-off-by: Abraham Jimenez <abraham.jimenez.romo@gmail.com>